# contracts/cloud/openapi.yaml
openapi: 3.1.0
info:
  title: AO Cloud Public API
  version: 1.0.0
  description: |
    Client-facing contract for a separate AO Cloud implementation. This
    repository does not serve these routes. The worker surface documents the
    exact outbound HTTP protocol used by sandbox workers, including one-time
    bootstrap, rotating credentials, fenced turn execution, orchestration, and
    durable workspace and terminal transport.
security:
  - bearerAuth: []
...
paths:
...
  /api/cloud/v1/orgs/{orgId}/projects:
    parameters:
      - $ref: "#/components/parameters/OrgId"
    get:
      operationId: listProjects
      tags: [Projects]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of projects in the tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectPage"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createProject
      tags: [Projects]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectInput"
      responses:
        "201":
          description: Project created.
          content:
            application/json:
              schema:
                type: object
                required: [project]
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
  /api/cloud/v1/orgs/{orgId}/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/ProjectId"
    patch:
      operationId: updateProject
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectInput"
      responses:
        "200":
          description: Project settings updated.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [project]
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteProject
      tags: [Projects]
      responses:
        "202":
          description: Project archived and sandbox teardown accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteProjectResponse"
        default:
          $ref: "#/components/responses/Error"
...
  /api/cloud/v1/orgs/{orgId}/sessions:
    parameters:
      - $ref: "#/components/parameters/OrgId"
    get:
      operationId: listSessions
      tags: [Sessions]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: projectId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: A page of sessions in the tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionPage"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createSession
      tags: [Sessions]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionInput"
      responses:
        "201":
          description: Session created.
          content:
            application/json:
              schema:
                type: object
                required: [session]
                properties:
                  session:
                    $ref: "#/components/schemas/Session"
        default:
          $ref: "#/components/responses/Error"
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/SessionId"
    get:
      operationId: getSession
      tags: [Sessions]
      responses:
        "200":
          description: Session details.
          content:
            application/json:
              schema:
                type: object
                required: [session]
                properties:
                  session:
                    $ref: "#/components/schemas/Session"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteSession
      tags: [Sessions]
      responses:
        "202":
          description: Session sandbox deletion was requested; durable history is retained.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSessionResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}/children:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/SessionId"
    get:
      operationId: listSessionChildren
      tags: [Sessions]
      description: >-
        Lists the sessions this orchestrator spawned, newest first, with each
        child's pull requests. Terminated children are included; a session
        that spawned nothing returns an empty page.
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of the orchestrator's child sessions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionChildrenPage"
        default:
          $ref: "#/components/responses/Error"
...
  /api/cloud/v1/orgs/{orgId}/provider-connections:
    parameters:
      - $ref: "#/components/parameters/OrgId"
    get:
      operationId: listProviderConnections
      tags: [Providers]
      responses:
        "200":
          description: Redacted provider connection metadata. Secrets are never returned.
          content:
            application/json:
              schema:
                type: object
                required: [providerConnections]
                properties:
                  providerConnections:
                    type: array
                    items:
                      $ref: "#/components/schemas/RedactedProviderConnection"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    workerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Session-scoped credential formatted as `Worker <token>`. Every
        authenticated route verifies both the token signature and that its
        worker ID and epoch are still current. Replacing a sandbox therefore
        fences an otherwise unexpired token with `STALE_WORKER_TOKEN`.

        Route scopes are `worker:connect` for heartbeat, `worker:event` for
        events, `worker:turn:claim`, `worker:turn:poll`, and
        `worker:turn:complete` for turn execution,
        `worker:credential:read` for the agent secret, `worker:git` for a
        checkout grant, `worker:orchestrate` for child controls, and
        `worker:transport` for workspace and terminal transport.
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Reusing a key with the same command returns the original result.
        Reusing it with a different command returns an IDEMPOTENCY_CONFLICT.
      schema:
        type: string
        minLength: 1
        maxLength: 200
  responses:
    Error:
      description: Standard Cloud API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    Project:
      type: object
      additionalProperties: false
      required: [id, orgId, displayName, repositoryUrl, defaultBranch, config, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        displayName:
          type: string
        repositoryUrl:
          type: string
          format: uri
        defaultBranch:
          type: string
        githubRepositoryId:
          type: string
          pattern: "^[1-9][0-9]*$"
          description: GitHub's integer repository ID encoded as a decimal string to preserve precision.
        config:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateProjectInput:
      type: object
      additionalProperties: false
      required: [displayName, repositoryUrl, defaultBranch]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 120
        repositoryUrl:
          type: string
          format: uri
          pattern: "^https://"
        defaultBranch:
          type: string
          minLength: 1
          maxLength: 255
        config:
          type: object
          additionalProperties: true
    UpdateProjectInput:
      type: object
      additionalProperties: false
      required: [displayName, defaultBranch]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 120
        defaultBranch:
          type: string
          minLength: 1
          maxLength: 255
    DeleteProjectResponse:
      type: object
      additionalProperties: false
      required: [project]
      properties:
        project:
          type: object
          additionalProperties: false
          required: [id, deleted]
          properties:
            id:
              type: string
              format: uuid
            deleted:
              type: boolean
    ProjectPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        page:
          $ref: "#/components/schemas/PageInfo"
...
    SessionKind:
      type: string
      enum: [worker, orchestrator]
...
    Session:
      type: object
      additionalProperties: false
      required:
        - id
        - orgId
        - projectId
        - kind
        - harness
        - displayName
        - branch
        - mode
        - deniedCommands
        - activityState
        - status
        - runtimeConnected
        - isTerminated
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        kind:
          $ref: "#/components/schemas/SessionKind"
        harness:
          type: string
        displayName:
          type: string
        branch:
          type: string
        mode:
          $ref: "#/components/schemas/SessionMode"
        deniedCommands:
          type: array
          items:
            type: string
        activityState:
          $ref: "#/components/schemas/SessionActivityState"
        status:
          $ref: "#/components/schemas/SessionStatus"
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/AgentCapability"
        runtimeConnected:
          type: boolean
        runtimeState:
          type: string
        runtimeError:
          type: string
        activeTurn:
          $ref: "#/components/schemas/Turn"
        isTerminated:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSessionInput:
      type: object
      additionalProperties: false
      required: [projectId, kind, harness, displayName, prompt]
      properties:
        projectId:
          type: string
          format: uuid
        kind:
          $ref: "#/components/schemas/SessionKind"
        harness:
          type: string
          minLength: 1
          maxLength: 120
        displayName:
          type: string
          minLength: 1
          maxLength: 80
        prompt:
          type: string
          maxLength: 65536
        mode:
          $ref: "#/components/schemas/SessionMode"
          default: trusted
        deniedCommands:
          type: array
          items:
            type: string
          default: []
        sandboxProviderConnectionId:
          type: string
          format: uuid
    SessionPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Session"
        page:
          $ref: "#/components/schemas/PageInfo"
...